refactor: move pose estimation helpers under utils
