+++ examples/pose-estimation/simple-pose-estimation-overlay-pipeline/cpp/utils/config.h
