+++ examples/pose-estimation/simple-pose-estimation-overlay-pipeline/python/tests/test_unit.py
@@ -42,10 +42,16 @@ class TestArgParsing:
         assert "unrecognized" in r.stderr.lower() or "error" in r.stderr.lower()
 
     def test_help_runs_without_sidecar_config_module(self):
-        """The packaged entrypoint should still show help when config.py is absent."""
+        """The packaged entrypoint should still show help with utils/config.py."""
         source = MAIN_PY.read_text(encoding="utf-8")
+        config_source = (
+            MAIN_PY.resolve().parent / "utils" / "config.py"
+        ).read_text(encoding="utf-8")
         with tempfile.TemporaryDirectory() as temp_dir:
             temp_main = Path(temp_dir) / "main.py"
+            temp_utils = Path(temp_dir) / "utils"
+            temp_utils.mkdir()
+            (temp_utils / "config.py").write_text(config_source, encoding="utf-8")
             temp_main.write_text(source, encoding="utf-8")
             r = subprocess.run(
                 [sys.executable, str(temp_main), "--help"],
