+++ examples/pose-estimation/simple-pose-estimation-overlay-pipeline/python/utils/config.py
